+++ lib/runners/processor/jshint.rb
@@ -25,6 +25,7 @@ module Runners
 
     def setup
       show_runtime_versions
+      add_warning_if_deprecated_options([:options], doc: "https://help.sider.review/tools/javascript/jshint")
       yield
     end