+++ test/smokes/misspell/expectations.rb
@@ -46,6 +46,12 @@ Smoke.add_test("locale", {
       name: 'Misspell',
       version: '0.3.4'
     }
+}, {
+  warnings: [{ message: <<~MSG.strip, file: "sideci.yml" }],
+    DEPRECATION WARNING!!!
+    The `$.linter.misspell.options` option(s) in your `sideci.yml` are deprecated and will be removed in the near future.
+    Please update to the new option(s) according to our documentation (see https://help.sider.review/tools/others/misspell ).
+  MSG
 })
 
 Smoke.add_test("ignore", {